refactor(hdac-pos): factor out delegations into a module
